backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
Pillow==11.1.0
opencv-python==4.11.0.86
pytesseract==0.3.13
numpy==2.0.2
pytest==8.0.0
pytest-cov==4.1.0
google-cloud-vision==3.5.0 